requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
djangorestframework==3.13.1
feedparser==6.0.8
Pillow==9.0.1
psycopg2-binary==2.9.3
pytz==2021.3
sgmllib3k==1.0.0
sqlparse==0.4.2
